+++ packages/container/libraries/core/src/planning/actions/addRootServiceNodeBindingIfContextFree.spec.ts
@@ -19,7 +19,7 @@ import { InternalBindingConstraints } from '../../binding/models/BindingConstrai
 import { bindingScopeValues } from '../../binding/models/BindingScope';
 import { bindingTypeValues } from '../../binding/models/BindingType';
 import { SingleInmutableLinkedList } from '../../common/models/SingleInmutableLinkedList';
-import { PlanServiceNodeBindingAddedResult } from '../../metadata/models/PlanServiceNodeUpdateResult';
+import { PlanServiceNodeBindingAddedResult } from '../../metadata/models/PlanServiceNodeAddedResult';
 import { buildPlanBindingConstraintsList } from '../calculations/buildPlanBindingConstraintsList';
 import { LazyPlanServiceNode } from '../models/LazyPlanServiceNode';
 import { PlanParams } from '../models/PlanParams';
